requirements.txt
# Lite SD-WAN Routing System Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.10.0

# Network Tools
ping3==4.0.4
networkx==3.2.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.3

# HTTP Client
requests==2.31.0

# Configuration
pyyaml==6.0.1
